simulation core refactor
